Make changes to Gameplay Scene

+++ src/Class/scenes/main_menu_scene.cpp
@@ -15,7 +15,7 @@ MainMenuScene::MainMenuScene()
 	gameVersionText.setString(gameVersion);
 	gameVersionText.setCharacterSize(15);
 	gameVersionText.setOrigin(gameVersionText.getLocalBounds().width, gameVersionText.getLocalBounds().height);
-	gameVersionText.setPosition(screenWidth - 15, screenHeight - 15);
+	gameVersionText.setPosition(static_cast<float>(screenWidth - 15), static_cast<float>(screenHeight - 15));
 
 	playButton = new Button(Vector2f(screenWidth * 0.5f, screenHeight * 0.5f), "JUGAR NIVEL" + to_string(gameLevel), 25);
 	playButton->setFont(font);
